PathFinding: dijkstra: accept directed graphs
The old method in 'buildGraph' forced all graphs to be undirected; this
change allows for directed graphs. This makes the algorithm more
flexible; if we want the old behavior, we can simply edit the edge list
to contain dual edges ('a' to 'b', and 'b' to 'a') before feeding it
into 'buildGraph'.

We also fix a bug where 'shortestPath' always included the destination
vertex, even if it was unreachable.